chore(model): RPC conversion code isolation (instance)
