+++ crates/api-model/src/instance/status/tenant.rs
@@ -15,7 +15,6 @@
  * limitations under the License.
  */
 
-use ::rpc::errors::RpcDataConversionError;
 use serde::{Deserialize, Serialize};
 
 /// The most recent tenant related status
@@ -25,17 +24,6 @@ pub struct InstanceTenantStatus {
     pub state: TenantState,
     /// An optional message which can contain details about the state
     pub state_details: String,
-}
-
-impl TryFrom<InstanceTenantStatus> for rpc::InstanceTenantStatus {
-    type Error = RpcDataConversionError;
-
-    fn try_from(state: InstanceTenantStatus) -> Result<Self, Self::Error> {
-        Ok(rpc::InstanceTenantStatus {
-            state: rpc::TenantState::try_from(state.state)? as i32,
-            state_details: state.state_details,
-        })
-    }
 }
 
 /// Enumerates possible instance states from the view of a tenant
@@ -71,25 +59,6 @@ pub enum TenantState {
     Failed,
     /// Not sure what happened. Check log for more info
     Invalid,
-}
-
-impl TryFrom<TenantState> for rpc::TenantState {
-    type Error = RpcDataConversionError;
-
-    fn try_from(state: TenantState) -> Result<Self, Self::Error> {
-        Ok(match state {
-            TenantState::Provisioning => rpc::TenantState::Provisioning,
-            TenantState::DpuReprovisioning => rpc::TenantState::DpuReprovisioning,
-            TenantState::Ready => rpc::TenantState::Ready,
-            TenantState::Configuring => rpc::TenantState::Configuring,
-            TenantState::Terminating => rpc::TenantState::Terminating,
-            TenantState::Terminated => rpc::TenantState::Terminated,
-            TenantState::Failed => rpc::TenantState::Failed,
-            TenantState::HostReprovisioning => rpc::TenantState::HostReprovisioning,
-            TenantState::Updating => rpc::TenantState::Updating,
-            TenantState::Invalid => rpc::TenantState::Invalid,
-        })
-    }
 }
 
 #[cfg(test)]
